EDIV/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-tinymce==3.7.1
python-decouple==3.8
python-magic==0.4.27
sqlparse==0.4.4
